requirements.txt
Django==5.2
django-environ==0.12.0
gunicorn
uvicorn
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf_yasg
redis
django-redis
psycopg2-binary
pillow
django-modeltranslation